feat: Update text content and innerHTML in DOM examples
This commit updates the text content and innerHTML in the DOM examples to improve clarity and accuracy. In "text-content.html", the text content of the <p> element with id "replace" is changed to "This text is replaced". In "innerHTML.html", the innerHTML of the <ul> element with id "list" is modified to include the correct closing tags for the list items. These changes ensure that the examples provide accurate and reliable information.